[MODIFY] deleteBoard시 img없을 경우 예외처리 추가(#236)

+++ src/api/boards/boards.service.ts
@@ -222,9 +222,11 @@ export class BoardsService {
       id: board.location.id,
     });
 
-    await this.imagesService.deleteImages({
-      url: board.image,
-    });
+    if (!board.image[0].url.includes('default.img')) {
+      await this.imagesService.deleteImages({
+        url: board.image,
+      });
+    }
 
     const result = await this.boardRepository.softDelete({ id: boardId });
